adicao do cadastro de transação e migrations de teste

+++ database/migrations/0001_01_01_000002_create_transactions_table.php
@@ -11,17 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('transactions', function (Blueprint $table) {
-            $table->id();
-            $table->decimal('amount', 10, 2)->nullable()->change();;
-            $table->unsignedBigInteger('category_id'); 
-            $table->timestamps();
-        
-            $table->foreign('category_id')
-                ->references('id')
-                ->on('categories')
-                ->onDelete('restrict'); 
+        if (!Schema::hasTable('transactions')){
+            Schema::create('transactions', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->string('tipo');
+                $table->decimal('valor', 10, 2);
+                $table->unsignedBigInteger('category_id');
+                $table->timestamps();
+            
+                $table->foreign('user_id')->references('id')->on('users');
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');;
+            
+                $table->foreign('category_id')
+                    ->references('id')
+                    ->on('categories')
+                    ->onDelete('restrict'); 
         });
-        
+    }
     }      
 };
